memory_service/requirements.txt
fastapi
uvicorn[standard]
watchdog
sentence-transformers
torch
numpy
pydantic
python-dotenv
pypdf
python-docx
openpyxl
python-pptx
pillow
pytesseract
sqlalchemy
pyyaml

